Split window code to different file
